[ci_cd] Minor scripting improvements

+++ ci_cd/live_cd/01-pre-chroot-script.sh
@@ -4,7 +4,7 @@ set -x
 DANGERZONE_VERSION=$1
 THIS_SCRIPTS_DIR="$(realpath $(dirname "$0"))"
 echo "Cleanup previous build"
-sudo rm -rf $HOME/LIVE_BOOT
+test -d $HOME/LIVE_BOOT && sudo rm -rf $HOME/LIVE_BOOT
 
 echo "Install required packages"
 sudo apt update && sudo apt install -y \
@@ -44,7 +44,7 @@ sudo cp /tmp/dangerzone_release $HOME/LIVE_BOOT/chroot/etc/dangerzone_release
 cp "${THIS_SCRIPTS_DIR}"/../../artifacts/dangerzone-linux*/dangerzone-cli /tmp/live-dangerzone-cli
 cp "${THIS_SCRIPTS_DIR}"/../../artifacts/dangerzone-linux*/dangerzone-httpserver /tmp/live-dangerzone-httpserver
 
-rm /tmp/live-dangerzone-container.tar
+test -f /tmp/live-dangerzone-container.tar && rm /tmp/live-dangerzone-container.tar
 podman build -f "${THIS_SCRIPTS_DIR}"/../../dangerzone_container/Dockerfile -t "docker.io/uycyjnzgntrn/dangerzone-converter:${DANGERZONE_VERSION}"
 podman save -o /tmp/live-dangerzone-container.tar "docker.io/uycyjnzgntrn/dangerzone-converter:${DANGERZONE_VERSION}"
