+++ ci_cd/live_cd/05-post-chroot-script.sh
@@ -5,8 +5,8 @@ ROOT_SCRIPTS_DIR="$(realpath $(dirname "$0"))"
 
 echo "Deleting previous artifacts ISO and squashfs files"
 DANGERZONE_VERSION=$(cat $HOME/LIVE_BOOT/chroot/etc/dangerzone_release | head -1)
-rm $HOME/LIVE_BOOT/dangerzone-livecd-amd64-${DANGERZONE_VERSION}.iso
-sudo rm $HOME/LIVE_BOOT/staging/live/filesystem.squashfs
+test -f $HOME/LIVE_BOOT/dangerzone-livecd-amd64-${DANGERZONE_VERSION}.iso && rm $HOME/LIVE_BOOT/dangerzone-livecd-amd64-${DANGERZONE_VERSION}.iso
+test -f $HOME/LIVE_BOOT/staging/live/filesystem.squashfs && sudo rm $HOME/LIVE_BOOT/staging/live/filesystem.squashfs
 
 echo "Creating filesystem"
 mkdir -p $HOME/LIVE_BOOT/staging/EFI/boot/
